Extract tests for isUint8ClampedArray to own file
